refactor: 대시보드 리팩토링
