+++ package.json
@@ -34,8 +34,6 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
-    "@types/lodash": "^4.14.182",
-    "bignumber": "^1.1.0",
     "bignumber.js": "^9.0.2",
     "classnames": "^2.3.1",
     "date-fns": "^2.28.0",
@@ -44,7 +42,6 @@
     "react": "^18.1.0",
     "react-datepicker": "^4.8.0",
     "react-dom": "^18.1.0",
-    "react-error-boundary": "^3.1.4",
     "react-query": "^3.39.0",
     "react-router-dom": "6",
     "react-scripts": "5.0.1",
@@ -56,6 +53,7 @@
   },
   "devDependencies": {
     "@types/jest": "^27.0.1",
+    "@types/lodash": "^4.14.182",
     "@types/node": "^16.7.13",
     "@types/react": "^18.0.0",
     "@types/react-datepicker": "^4.4.1",
