+++ src/routes/Dashboard/TotalAdsChart/totalAdsChart.module.scss
@@ -1,0 +1,14 @@
+.dropdownContainer {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 54px;
+  margin-bottom: 40px;
+
+  .filterDropdown {
+    display: flex;
+
+    div {
+      margin-right: 10px;
+    }
+  }
+}
