+++ src/routes/Dashboard/model.ts
