+++ src/routes/Dashboard/utils/chartUtils.ts
@@ -1,38 +1,34 @@
 import dayjs from 'dayjs'
+
 import { IDaily, IChart, IMenu } from 'types/trend'
-import BigNumber from 'bignumber.js'
-import { getPlus } from 'utils/num'
+import { getPlus, Num } from 'utils/num'
 
-const Num = (n: string | number, b?: number | undefined): BigNumber => {
-  if (typeof n === 'string') {
-    return new BigNumber(n.replace(/,/g, ''), b)
-  }
-
-  return new BigNumber(n, b)
-}
-
-export const getDays = ([startDate, stopDate]: string[]): string[] => {
-  const dateArray = []
+const getDays = ([startDate, stopDate]: string[]): string[] => {
   let currentDate = dayjs(startDate)
   const lastDate = dayjs(stopDate)
+  const dateArray = []
+
   while (currentDate <= lastDate) {
     dateArray.push(dayjs(currentDate).format('YYYY-MM-DD'))
     currentDate = dayjs(currentDate).add(1, 'days')
   }
+
   return dateArray
 }
 
-export const getDates = ([startDate, stopDate]: string[]): dayjs.Dayjs[] => {
+const getDates = ([startDate, stopDate]: string[]): dayjs.Dayjs[] => {
   const dateArray = []
 
   if (stopDate === undefined) return [dayjs(stopDate)]
 
   let currentDate = dayjs(startDate)
   const lastDate = dayjs(stopDate)
+
   while (currentDate <= lastDate) {
     dateArray.push(dayjs(currentDate))
     currentDate = dayjs(currentDate).add(1, 'days')
   }
+
   return dateArray
 }
 
@@ -54,25 +50,30 @@ const findTargetValue = (target: string, obj: IDaily) => {
   return targetObejct
 }
 
-export const getChartData = (selectedDate: any[], rowChartData: IDaily[], target: string) => {
+const getChartData = (selectedDate: any[], rowChartData: IDaily[], target: string) => {
   const data: IChart[] = []
+
   if (target === 'none') return []
+
   selectedDate.forEach((date) => {
     rowChartData.some((obj) => {
       if (obj.date === date) {
         const value = findTargetValue(target, obj)
         data.push({ x: dayjs(date), y: value })
       }
+
       return obj.date === date
     })
   })
+
   return data
 }
 
-export const ChangeText = (value: number, type: string) => {
+const getChangeText = (value: number, type: string) => {
   const countType = ['click', 'conv', 'imp']
   const monenyType = ['convValue', 'cost', 'cpc', 'cpa']
   let reulst = ''
+
   if (countType.includes(type)) {
     reulst = `${Math.round(value).toLocaleString()}번`
   }
@@ -92,13 +93,15 @@ export const ChangeText = (value: number, type: string) => {
   return reulst
 }
 
-export const isMenu = (element: IMenu, value: string) => {
+const isMenu = (element: IMenu, value: string) => {
   if (element.value === value) return false
+
   return true
 }
 
-export const ChartYsum = (obj: IChart[]) => {
+const sumChartY = (obj: IChart[]) => {
   if (obj.length === 0) return -9999
+
   const sum = obj
     .map((item) => {
       return item.y
@@ -107,3 +110,5 @@ export const ChartYsum = (obj: IChart[]) => {
 
   return sum
 }
+
+export { getDays, getDates, getChartData, getChangeText, isMenu, sumChartY }
